+++ CurveBase/CurveData/CurveInterpolatedData/LagarangePolynomialCurveInterpolatedData.cs
@@ -20,36 +20,32 @@ using CurveBase.CurveData.CurveParam;
 using CurveBase.CurveElement.IntervalPolynomialCurve;
 using Util.Tool;
 using Util.Variable;
+using Util.Variable.Interval;
+using Util.Variable.Matrix;
 using Util.Variable.PointList;
 
 namespace CurveBase.CurveData.CurveInterpolatedData
 {
     public class LagarangePolynomialCurveInterpolatedData : ICurveInterpolatedData
     {
-        private PiecewiseIntervalPolynomialCurveElement curve = null;
+        private NormalIntervalPolynomialCurveElement curve = null;
         private OrderedCurvePointList lagarangePolynomialInterpolatedPoints = null;
         private PolynomialCurveType curveType;
         private DataPoint lastPoint = null;
+        private int count = 0;
 
         #region Constructor
         public LagarangePolynomialCurveInterpolatedData(PolynomialCurveParam curveParam)
         {
             this.curveType = curveParam.PolynomialCurveType;
             this.lastPoint = curveParam.PointList.RightBorderPoint;
-            switch (curveType)
-            {
-                case PolynomialCurveType.Lagrange_Linear:
-                    lagarangePolynomialInterpolatedPoints = curveParam.PointList;
-                    break;
-                case PolynomialCurveType.Lagrange_Quadratic:
-                    curve = generateQuadraticPolynomialCurves(curveParam);
-                    break;
-            }
+            this.count = curveParam.Count;
+            this.curve = GenerateCurve(curveParam);
         }
         #endregion
 
         #region Property
-        public PiecewiseIntervalPolynomialCurveElement Curves
+        public NormalIntervalPolynomialCurveElement Curve
         {
             get
             {
@@ -62,16 +58,6 @@ namespace CurveBase.CurveData.CurveInterpolatedData
             get
             {
                 return lagarangePolynomialInterpolatedPoints;
-            }
-        }
-
-        public int Count
-        {
-            get
-            {
-                if (curve != null) return curve.Curves.Count;
-                if (lagarangePolynomialInterpolatedPoints != null) return lagarangePolynomialInterpolatedPoints.Count - 1;
-                throw new ArgumentException("The LagarangePolynomialInterpolatedData hasn't been initialized properly.");
             }
         }
         #endregion
@@ -89,21 +75,46 @@ namespace CurveBase.CurveData.CurveInterpolatedData
         #endregion
 
         #region Private.Methods
-        private PiecewiseIntervalPolynomialCurveElement generateQuadraticPolynomialCurves(PolynomialCurveParam param)
+        private NormalIntervalPolynomialCurveElement GenerateCurve(PolynomialCurveParam curveParam)
         {
-            Debug.Assert(param.PolynomialCurveType == PolynomialCurveType.Lagrange_Quadratic, @"This method """"generateQuadraticPolynomialCurves"""" only supports quadratic polynomialCurveType");
-            OrderedCurvePointList pointList = param.PointList;
-            List<NormalIntervalPolynomialCurveElement> polynomialCurve = new List<NormalIntervalPolynomialCurveElement>();
-            List<DoubleExtension> coefficients = null;
-            List<DoubleExtension> cutPoints = new List<DoubleExtension>();
-            cutPoints.Add(pointList[0].X);
-            for (int i = 2; i < pointList.Count; i = i + 2)
+            Debug.Assert(curveParam.PolynomialCurveType == PolynomialCurveType.Lagrange, @"This interpolated method """"GenerateCurve"""" only supports Lagarange Polynomial Curve.");
+            double[,] coefficientsArray = new double[count, count];
+            double[,] constantArray = new double[count, 1];
+            double coef, x;
+            
+            for (int i = 0; i < count; i++)
             {
-                coefficients = MathExtension.CalculateQuadraticPolynomialCoefficients(pointList[i - 2], pointList[i - 1], pointList[i]);
-                polynomialCurve.Add(new NormalIntervalPolynomialCurveElement(coefficients, 3, pointList[i - 2].X, pointList[i].X));
-                cutPoints.Add(pointList[i].X);
+                constantArray[i, 0] = curveParam[i].Y.AccurateValue;
+                coef = 1;
+                x = curveParam[i].X.AccurateValue;
+                for (int j = 0; j < count; j++)
+                {
+                    coefficientsArray[i, j] = coef;
+                    coef *= x;
+                }
             }
-            return new PiecewiseIntervalPolynomialCurveElement(polynomialCurve, cutPoints);
+
+            LinearEquationSet equations = new LinearEquationSet(new Matrix(coefficientsArray), new Matrix(constantArray));
+            Matrix result = equations.AnswerMatrix;
+
+            List<DoubleExtension> coefs = new List<DoubleExtension>();
+            for (int i = 0; i < count; i++)
+            {
+                coefs.Add(new DoubleExtension(result[i, 0]));
+            }
+            return new NormalIntervalPolynomialCurveElement(coefs, count, new DataInterval(curveParam.PointList.LeftBorderPoint.X, curveParam.PointList.RightBorderPoint.X)); 
+            //OrderedCurvePointList pointList = param.PointList;
+            //List<NormalIntervalPolynomialCurveElement> polynomialCurve = new List<NormalIntervalPolynomialCurveElement>();
+            //List<DoubleExtension> coefficients = null;
+            //List<DoubleExtension> cutPoints = new List<DoubleExtension>();
+            //cutPoints.Add(pointList[0].X);
+            //for (int i = 2; i < pointList.Count; i = i + 2)
+            //{
+            //    coefficients = MathExtension.CalculateQuadraticPolynomialCoefficients(pointList[i - 2], pointList[i - 1], pointList[i]);
+            //    polynomialCurve.Add(new NormalIntervalPolynomialCurveElement(coefficients, 3, pointList[i - 2].X, pointList[i].X));
+            //    cutPoints.Add(pointList[i].X);
+            //}
+            //return new PiecewiseIntervalPolynomialCurveElement(polynomialCurve, cutPoints);
         }
         #endregion
     }